readd support for shorthands from /shorthand/store
