+++ packages/store/ts/config/v2/index.ts
@@ -1,7 +1,7 @@
 export * from "./generics";
 export * from "./scope";
 export * from "./schema";
-export * from "./shorthand/tableShorthand";
+export * from "./shorthand/table";
 export * from "./table";
 export * from "./store";
 export * from "./output";
